+++ Source/ImoutoRebirth.Room/ImoutoRebirth.Room.Application/Cqrs/FoldersSlice/DeleteSourceFolderCommand.cs
@@ -9,17 +9,22 @@ public record DeleteSourceFolderCommand(Guid CollectionId, Guid SourceFolderId) 
 internal class DeleteSourceFolderCommandHandler : ICommandHandler<DeleteSourceFolderCommand>
 {
     private readonly ICollectionRepository _collectionRepository;
+    private readonly IEventStorage _eventStorage;
 
-    public DeleteSourceFolderCommandHandler(ICollectionRepository collectionFileRepository) 
-        => _collectionRepository = collectionFileRepository;
+    public DeleteSourceFolderCommandHandler(ICollectionRepository collectionFileRepository, IEventStorage eventStorage)
+    {
+        _collectionRepository = collectionFileRepository;
+        _eventStorage = eventStorage;
+    }
 
     public async Task Handle(DeleteSourceFolderCommand request, CancellationToken cancellationToken)
     {
         var (collectionId, sourceFolderId) = request;
         var collection = await _collectionRepository.GetById(collectionId).GetAggregateOrThrow();
 
-        collection.RemoveSourceFolder(sourceFolderId);
+        var result = collection.RemoveSourceFolder(sourceFolderId);
 
+        _eventStorage.AddRange(result.EventsCollection);
         await _collectionRepository.Update(collection);
     }
 }