+++ Source/ImoutoRebirth.Room/ImoutoRebirth.Room.UI/Scheduling/FileSystem/FileSystemWatcherEvent.cs
@@ -1,3 +1,0 @@
-﻿namespace ImoutoRebirth.Room.UI.Scheduling.FileSystem;
-
-public record FileSystemWatcherEvent(string Folder, FileSystemWatcherEventType Type, FileSystemEventArgs Args);