+++ Source/ImoutoRebirth.Room/ImoutoRebirth.Room.UI/Scheduling/FileSystem/FileSystemWatcherEventType.cs
@@ -1,9 +1,0 @@
-﻿namespace ImoutoRebirth.Room.UI.Scheduling.FileSystem;
-
-public enum FileSystemWatcherEventType
-{
-    Changed,
-    Created,
-    Deleted,
-    Renamed
-}