+++ Source/ImoutoRebirth.Room/ImoutoRebirth.Room.UI/ServiceCollectionExtensions.cs
@@ -1,5 +1,5 @@
-﻿using ImoutoRebirth.Room.UI.Scheduling;
-using ImoutoRebirth.Room.UI.Scheduling.FileSystem;
+﻿using ImoutoRebirth.Room.Application.Services;
+using ImoutoRebirth.Room.UI.FileSystemEvents;
 using Microsoft.Extensions.DependencyInjection;
 
 namespace ImoutoRebirth.Room.UI;
@@ -8,10 +8,6 @@ public static class ServiceCollectionExtensions
 {
     public static IServiceCollection AddRoomUi(this IServiceCollection services)
     {
-        // services.Configure<PeriodicRunnerOptions>(x => x.AddJob<OverseeJob>());
-        // services.AddTransient<OverseeJob>();
-        // services.AddHostedService<PeriodicRunnerHostedService>();
-
         services.AddHostedService<SourceFolderWatcherHostedService>();
 
         services.AddSingleton<SourceFolderWatcher>();